refactor(timesync): use single-precision exp for filter gain schedule
The gain-schedule sigmoid was the only double-precision exp() user in
the entire firmware, keeping libm's exp and its 2.1 KB __exp_data table
linked. Float precision is more than sufficient for a blend factor
between 0.05 and 0.003; the filter state itself remains double.

Saves 2768 bytes of flash on px4_fmu-v6x_default.

+++ src/lib/timesync/Timesync.cpp
@@ -75,9 +75,10 @@ void Timesync::update(const uint64_t now_us, const int64_t remote_timestamp_ns, 
 
 				// Filter gain scheduling
 				if (!sync_converged()) {
-					// Interpolate with a sigmoid function
-					double progress = (double)_sequence / (double)CONVERGENCE_WINDOW;
-					double p = 1.0 - exp(0.5 * (1.0 - 1.0 / (1.0 - progress)));
+					// Interpolate with a sigmoid function (float precision is sufficient
+					// for the gain schedule and avoids linking double-precision exp())
+					float progress = (float)_sequence / (float)CONVERGENCE_WINDOW;
+					double p = (double)(1.f - expf(0.5f * (1.f - 1.f / (1.f - progress))));
 					_filter_alpha = p * ALPHA_GAIN_FINAL + (1.0 - p) * ALPHA_GAIN_INITIAL;
 					_filter_beta = p * BETA_GAIN_FINAL + (1.0 - p) * BETA_GAIN_INITIAL;
